pyproject.toml
[tool.poetry]
name = "tatarubot2"
version = "0.1.6"
description = "FF14 bot Tataru"
authors = ["aaron-li"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7.3"
nonebot2 = "^2.0.0-beta.1"
requests = "*"

[tool.poetry.dev-dependencies]
nb-cli = "^0.6.0"

[tool.nonebot]
plugins = []
plugin_dirs = ["tatarubot2/plugins"]

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
